Allow usage with ZF3 applications
This patch removes collection of version data from the plugin, as zend-version
is no longer incluced with ZF3, and has not been updated past the 2.4 series.
Currently, if you use this with ZF3, a fatal error occurs as zend-version
is not installed.

The patch also replaces the logo with one that removes the version number, as we
have re-branded to remove versioning from the logo within the project.